+++ game/test/view/test_view.py
@@ -3,6 +3,7 @@ import sys
 import os
 sys.path.append(os.path.abspath('../../'))
 from view import view
+from controller import controller
 import pygame
 
 
@@ -42,6 +43,36 @@ class TestView(unittest.TestCase):
         # assert expected outcome
         for result in results:
             self.assertIsNotNone(result)
+    
+    def test_draw_settings_screen(self):
+        # set object state
+        self.test_view.set_up_fonts()
+        # execute method
+        results = self.test_view.draw_settings_screen("player_name", "host_ip")
+        # assert expected outcome
+        for result in results:
+            self.assertIsNotNone(result)
+    
+    def test_animate_capture(self):
+        # set object state
+        ray_coords = dict()
+        ray_coords['top'] = 100
+        ray_coords['x'] = 10
+        ray_coords['y'] = 200
+        ray_coords['c'] = 10
+        controller.Controller.load_external_resources(self.test_view, os.path.join(os.getcwd(), '../../'))
+        
+        # execute method
+        outcome = self.test_view.animate_capture(ray_coords)
+        # assert expected outcome
+        self.assertTrue(outcome)
+        
+        # reset data
+        ray_coords['y'] = 10
+        # execute method
+        outcome = self.test_view.animate_capture(ray_coords)
+        # assert expected outcome
+        self.assertFalse(outcome)
         
 
 if __name__ == '__main__':